api_server/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
websockets==12.0
python-dotenv==1.0.0
google-genai
passlib[bcrypt]==1.7.4
pyjwt==2.8.1
python-multipart==0.0.6
psycopg2-binary==2.9.9
pg8000==1.29.8
cloud-sql-python-connector==1.4.3
gunicorn==21.2.0
requests==2.31.0
httpx==0.26.0
python-jose[cryptography]==3.3.0
email-validator==2.1.0
pytz==2024.1
python-dateutil==2.8.2
